#188 #193 Landing page refactoring, donate feature, styling
