+++ .eslintrc.json
@@ -59,11 +59,13 @@
         "no-debugger": "warn",
         "react/jsx-uses-react": 2,
         "react/jsx-uses-vars": 2,
+        "react/no-unescaped-entities": 0,
         "react/react-in-jsx-scope": 2,
         "react/prop-types": 0,
         "react/destructuring-assignment": 0,
         "react/jsx-first-prop-new-line": 0,
         "react/jsx-filename-extension": 0,
+        "react/prefer-stateless-function": 0,
         "jsx-a11y/click-events-have-key-events": 0,
         "jsx-a11y/no-noninteractive-element-interactions": 0,
         "react/jsx-one-expression-per-line": 0
