+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 yarn-error.log
 .DS_Store
 env.js
+.idea
