+++ src/components/about.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import '../style.scss';
+
+class About extends Component {
+  render() {
+    return (
+      <div>
+        About Us
+      </div>
+    );
+  }
+}
+
+
+export default About;
