+++ src/components/app.js
@@ -4,6 +4,7 @@
 /* eslint-disable max-len */
 /* eslint-disable react/no-unescaped-entities */
 /* eslint-disable no-unused-vars */
+
 import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
 import {
@@ -13,1772 +14,40 @@ import '../style.scss';
 import AOS from 'aos';
 import MapView from './mapview';
 import Create from './create';
-import 'aos/dist/aos.css'; // You can also use <link> for styles
-// ..
+import 'aos/dist/aos.css';
+import Nav from '../containers/nav';
+import Footer from '../containers/footer';
+import Landing from './Landing';
+import Legal from './Legal';
+import About from './About';
+
 AOS.init(); // {mirror: true} changes the animations significantly but I think it goes a bit over the top
-
-
-export const navStyle = {
-  flexDirection: 'row', display: 'flex', flexWrap: 'wrap', justifyContent: 'space-evenly', flexFlow: 'row', listStyleType: 'none',
-};
-
-export const navOptionStyleHover = {
-  color: 'orange',
-  fontSize: '20px',
-};
-
-export const navOptionStyle = {
-  color: '#FFFFFF',
-};
-
-const Nav = (props) => {
-  const [isHover, setIsHover] = useState(false);
-  const [isHover1, setIsHover1] = useState(false);
-  const [isHover2, setIsHover2] = useState(false);
-
-
-  // @bella also change this. The <ol> / <li> should be changed; only thing to be preserved here exactly is the NavLink components
-  return (
-    <nav className="navBar">
-      <ol style={navStyle}>
-        <li style={isHover ? navOptionStyleHover : navOptionStyle}
-          onMouseEnter={() => { setIsHover(true); }}
-          onMouseLeave={() => { setIsHover(false); }}
-        >
-          <NavLink to="/" exact style={{ color: '#ff5722', fontFamily: 'Montserrat', fontWeight: 'bold' }}>Home</NavLink>
-        </li>
-        <li style={isHover1 ? navOptionStyleHover : navOptionStyle}
-          onMouseEnter={() => { setIsHover1(true); }}
-          onMouseLeave={() => { setIsHover1(false); }}
-        >
-          <NavLink to="/whatsup" style={{ color: '#ff5722', fontFamily: 'Montserrat' }}>Explore</NavLink>
-        </li>
-        <li style={isHover2 ? navOptionStyleHover : navOptionStyle}
-          onMouseEnter={() => { setIsHover2(true); }}
-          onMouseLeave={() => { setIsHover2(false); }}
-        >
-          <NavLink to="/create" style={{ color: '#ff5722', fontFamily: 'Montserrat' }}>Create portal </NavLink>
-        </li>
-        <li>
-          <NavLink to="/legal" style={{ color: '#ff5722', fontFamily: 'Montserrat' }}>Legal </NavLink>
-        </li>
-      </ol>
-
-    </nav>
-
-  );
-};
 
 const FallBack = (props) => {
   // this is our 404 page, this is also a low priority
   return <div>URL Not Found</div>;
 };
 
-
-const Welcome = (props) => {
-  // @bella the majority of your work will go right here in the return block of this function, this is your highest priority
-
-  const styleObj = {
-    fontSize: 25,
-    fontWeight: 'bold',
-    color: '#B5D8CF',
-    paddingTop: '100px',
-    backgroundColor: 'black',
-  };
-
-
-  // ^now you can re-use this style object where you need to
-
-  const boldText = {
-    fontFamily: 'Montserrat, sans-serif',
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: 'black',
-    justifyContent: 'center',
-  };
-
-  const blueHeaderStyle = {
-    fontFamily: 'Montserrat, sans-serif',
-    fontSize: 25,
-    fontWeight: 'bold',
-    color: '#005975',
-    paddingTop: '100px',
-  };
-
-  const blueBodyStyle = {
-    fontFamily: 'Montserrat, sans-serif',
-    fontSize: 12,
-    fontWeight: 'extra-light',
-    color: '#005975',
-  };
-
-  return (
-    <div>
-      <div className="titleBox" data-aos="zoom-out-down" data-aos-duration="2000">
-        <h1 style={{
-          fontFamily: 'Pacifico', color: 'white', display: 'flex', alignItems: 'center', justifyContent: 'center', marginTop: 40, marginBottom: 25, fontSize: 40,
-        }}
-        >Who's Down
-        </h1>
-        <h1 style={{
-          color: 'white', display: 'flex', alignItems: 'center', justifyContent: 'center', fontFamily: 'Open Sans', fontWeight: 'light', marginBottom: 60, marginLeft: 10, fontSize: 30,
-        }}
-        > Spontaneity Organized.
-        </h1>
-      </div>
-
-      <div className="mainPage">
-
-        <div className="imagine">Imagine a Campus Where...</div> {/* this style is ugly don't use this just wanted to show you an example */}
-        <div className="topBox" data-aos="fade-in" data-aos-duration="2000">
-          <div className="reasonBox">
-            <img className="icon" src="https://live.staticflickr.com/65535/50211873491_d76bacba51_q.jpg" alt="No_Commitment_Necessary_Icon" />
-            <div className="contentBottom"> </div>
-            <p className="basicText">No commitment is necessary to meet new people.</p>
-          </div>
-          <div className="reasonBox">
-            <img className="icon" src="https://live.staticflickr.com/65535/50211873611_bc226cd5e0_q.jpg" alt="User_Friendly_Interface_Icon" />
-            <div className="contentBottom"> </div>
-            <p className="basicText">A user-friendly interface shows what’s happening on campus right now.</p>
-          </div>
-          <div className="reasonBox">
-            <img className="icon" src="https://live.staticflickr.com/65535/50211346628_3040cda11e_q.jpg" alt="People_Not_Organizations_Icon" />
-            <div className="contentBottom"> </div>
-            <p className="basicText">People, not organizations, create events.</p>
-          </div>
-        </div>
-
-        <div>
-          <div className="imageSpacer"> </div>
-          <img className="image" src="https://live.staticflickr.com/65535/50201520242_9bdce03c13_b.jpg" alt="whosdowninfographic" />
-          <div className="imageSpacer"> </div>
-        </div>
-
-        <div className="appDesc" data-aos="zoom-in">
-          <div className="topBox"> <p className="basicTextBoldWhite">The meaningful connections of the ‘24s don’t have to be another casualty of COVID-19.</p></div>
-          <div className="topBox">
-            <p className="basicTextWhite">
-              Featuring hard event capacity limits that prevent discovery of full events,
-              rigorous data analytics and event moderation to measure and enforce compliance,
-              in-app reminders to follow public safety measures, and built in functionality
-              to organize real-life and virtual events alike, Who’s Down offers a way for ‘24s and
-              other Dartmouth undergraduates to ameliorate the pains of social distancing in a measurably safe way.
-            </p>
-          </div>
-        </div>
-
-        <div className="topBox">
-
-          <div className="reasonBox" data-aos="fade-left">
-            <p className="basicTextBold">Geographically and Systematically View Events Near You</p>
-
-            <p className="basicText">We help event seekers discover spontaneous events through geographic display and extensive filtering capabilities.</p>
-
-          </div>
-
-          <div className="reasonBox" data-aos="fade-in">
-            <p className="basicTextBold">Connect Students Across Communities</p>
-
-            <p className="basicText">We help you publish events  to create and plan them as they occur.</p>
-
-          </div>
-
-          <div className="reasonBox" data-aos="fade-right">
-            <p className="basicTextBold">Manage Remote Events</p>
-
-            <p className="basicText">We support in-person and virtual events, helping students keep track of Zoom links and in-person gatherings to connect with the events that matter to them.</p>
-
-          </div>
-
-          <div>
-            <div />
-          </div>
-
-
-        </div>
-      </div>
-    </div>
-  );
-};
-
-// ==================== EVERYTHING BELOW THIS LINE BELLA DOESN'T NEED TO TOUCH ==================== //
-
-const About = (props) => {
-  // don't touch this
+// TODO: Migrate Explore to new container when beginning to work on this
+const Explore = (props) => {
   return <MapView />;
 };
 
-const Legal = (props) => {
-  return (
-    <div>
-      <div><div align="center">
-        <div align="center">
-          <div align="center">
-            <div align="center">
-              <div align="center">
-                <div align="center">
-                  TERMS OF USE
-                </div>
-                <div align="center">
-                  <a name="_7m5b3xg56u7y" />
-                </div>
-                <div align="center">
-                  <br />
-                </div>
-                <div align="center">
-                  Last updated August 08, 2020
-                </div>
-              </div>
-            </div>
-            <div align="center">
-              <br />
-            </div>
-            <div align="center">
-              <br />
-            </div>
-            <div align="center">
-              <br />
-            </div>
-          </div>
-          <div>
-            <a name="_a7mwfgcrtsqn" />
-            AGREEMENT TO TERMS
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            These Terms of Use constitute a legally binding agreement made
-            between you, whether personally or on behalf of an entity
-            (&#8220;you&#8221;) and Who's Down LLC ("Company",
-            &#8220;we&#8221;, &#8220;us&#8221;, or &#8220;our&#8221;),
-            concerning your access to and use of the downatdart.me website as
-            well as any other media form, media channel, mobile website or
-            mobile application related, linked, or otherwise connected thereto
-            (collectively, the &#8220;Site&#8221;). You agree that by accessing
-            the Site, you have read, understood, and agreed to be bound by all
-            of these Terms of Use. IF YOU DO NOT AGREE WITH ALL OF THESE TERMS
-            OF USE, THEN YOU ARE EXPRESSLY PROHIBITED FROM USING THE SITE AND
-            YOU MUST DISCONTINUE USE IMMEDIATELY.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            Supplemental terms and conditions or documents that may be posted
-            on the Site from time to time are hereby expressly incorporated
-            herein by reference. We reserve the right, in our sole discretion,
-            to make changes or modifications to these Terms of Use at any time
-            and for any reason. We will alert you about any changes by updating
-            the &#8220;Last updated&#8221; date of these Terms of Use, and you
-            waive any right to receive specific notice of each such change. It
-            is your responsibility to periodically review these Terms of Use to
-            stay informed of updates. You will be subject to, and will be
-            deemed to have been made aware of and to have accepted, the changes
-            in any revised Terms of Use by your continued use of the Site after
-            the date such revised Terms of Use are posted.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            The information provided on the Site is not intended for
-            distribution to or use by any person or entity in any jurisdiction
-            or country where such distribution or use would be contrary to law
-            or regulation or which would subject us to any registration
-            requirement within such jurisdiction or country. Accordingly, those
-            persons who choose to access the Site from other locations do so on
-            their own initiative and are solely responsible for compliance with
-            local laws, if and to the extent local laws are applicable.
-          </div>
-          <div />
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            The Site is not tailored to comply with industry-specific
-            regulations (Health Insurance Portability and Accountability Act
-            (HIPAA), Federal Information Security Management Act (FISMA),
-            etc.), so if your interactions would be subjected to such laws, you
-            may not use this Site. You may not use the Site in a way that would
-            violate the Gramm-Leach-Bliley Act (GLBA).
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            The Site is intended for users who are at least 18 years old.
-            Persons under the age of 18 are not permitted to use or register
-            for the Site.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            <br />
-          </div>
-          <div>
-            <a name="_4rd71iod99ud" />
-            INTELLECTUAL PROPERTY RIGHTS
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            Unless otherwise indicated, the Site is our proprietary property
-            and all source code, databases, functionality, software, website
-            designs, audio, video, text, photographs, and graphics on the Site
-            (collectively, the &#8220;Content&#8221;) and the trademarks,
-            service marks, and logos contained therein (the
-            &#8220;Marks&#8221;) are owned or controlled by us or licensed to
-            us, and are protected by copyright and trademark laws and various
-            other intellectual property rights and unfair competition laws of
-            the United States, international copyright laws, and international
-            conventions. The Content and the Marks are provided on the Site
-            &#8220;AS IS&#8221; for your information and personal use only.
-            Except as expressly provided in these Terms of Use, no part of the
-            Site and no Content or Marks may be copied, reproduced, aggregated,
-            republished, uploaded, posted, publicly displayed, encoded,
-            translated, transmitted, distributed, sold, licensed, or otherwise
-            exploited for any commercial purpose whatsoever, without our
-            express prior written permission.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            Provided that you are eligible to use the Site, you are granted a
-            limited license to access and use the Site and to download or print
-            a copy of any portion of the Content to which you have properly
-            gained access solely for your personal, non-commercial use. We
-            reserve all rights not expressly granted to you in and to the Site,
-            the Content and the Marks.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            <br />
-          </div>
-          <div>
-            <a name="_vhkegautf00d" />
-            USER REPRESENTATIONS
-          </div>
-          <div>
-            <br />
-          </div>
-        </div>
-        <div align="center">
-          <div>
-            <div>
-              By using the Site, you represent and warrant that: (1) all
-              registration information you submit will be true, accurate,
-              current, and complete; (2) you will maintain the accuracy of
-              such information and promptly update such registration
-              information as necessary; (3) you have the legal capacity and
-              you agree to comply with these Terms of Use; (4) you are not a
-              minor in the jurisdiction in which you reside; (5) you will not
-              access the Site through automated or non-human means, whether
-              through a bot, script, or otherwise; (6) you will not use the
-              Site for any illegal or unauthorized purpose; and (7) your use
-              of the Site will not violate any applicable law or regulation.
-            </div>
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            <div>
-              If you provide any information that is untrue, inaccurate, not
-              current, or incomplete, we have the right to suspend or
-              terminate your account and refuse any and all current or future
-              use of the Site (or any portion thereof).
-            </div>
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            <br />
-          </div>
-          <div>
-            <a name="_esuoutkhaf53" />
-            <div>
-              USER REGISTRATION
-            </div>
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            <div>
-              You may be required to register with the Site. You agree to
-              keep your password confidential and will be responsible for all
-              use of your account and password. We reserve the right to
-              remove, reclaim, or change a username you select if we
-              determine, in our sole discretion, that such username is
-              inappropriate, obscene, or otherwise objectionable.
-            </div>
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div />
-          <div>
-            <a name="_1voziltdxegg" />
-            <div>
-              PROHIBITED ACTIVITIES
-            </div>
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            <div>
-              You may not access or use the Site for any purpose other than
-              that for which we make the Site available. The Site may not be
-              used in connection with any commercial endeavors except those
-              that are specifically endorsed or approved by us.
-            </div>
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            <div>
-              <div>
-                <div>
-                  As a user of the Site, you agree not to:
-                </div>
-                <div>
-                  <div>
-                    <br />
-                  </div>
-                </div>
-              </div>
-              <div>
-                1. Systematically retrieve data or other content from the
-                Site to create or compile, directly or indirectly, a
-                collection, compilation, database, or directory without
-                written permission from us.
-              </div>
-              <div>
-                2. Trick, defraud, or mislead us and other users,
-                especially in any attempt to learn sensitive account
-                information such as user passwords.
-              </div>
-              <div>
-                3. Circumvent, disable, or otherwise interfere with
-                security-related features of the Site, including features
-                that prevent or restrict the use or copying of any Content
-                or enforce limitations on the use of the Site and/or the
-                Content contained therein.
-              </div>
-              <div>
-                4. Disparage, tarnish, or otherwise harm, in our opinion,
-                us and/or the Site.
-              </div>
-              <div>
-                5. Use any information obtained from the Site in order to
-                harass, abuse, or harm another person.
-              </div>
-              <div>
-                6. Make improper use of our support services or submit
-                false reports of abuse or misconduct.
-              </div>
-              <div>
-                7. Use the Site in a manner inconsistent with any
-                applicable laws or regulations.
-              </div>
-              <div>
-                8. Use the Site to advertise or offer to sell goods and
-                services.
-              </div>
-              <div>
-                9. Engage in unauthorized framing of or linking to the
-                Site.
-              </div>
-              <div>
-                10. Upload or transmit (or attempt to upload or to
-                transmit) viruses, Trojan horses, or other material,
-                including excessive use of capital letters and spamming
-                (continuous posting of repetitive text), that interferes
-                with any party&#8217;s uninterrupted use and enjoyment of
-                the Site or modifies, impairs, disrupts, alters, or
-                interferes with the use, features, functions, operation, or
-                maintenance of the Site.
-              </div>
-              <div>
-                11. Engage in any automated use of the system, such as
-                using scripts to send comments or messages, or using any
-                data mining, robots, or similar data gathering and
-                extraction tools.
-              </div>
-              <div>
-                12. Delete the copyright or other proprietary rights notice
-                from any Content.
-              </div>
-              <div>
-                13. Attempt to impersonate another user or person or use
-                the username of another user.
-              </div>
-              <div>
-                14. Sell or otherwise transfer your profile.
-              </div>
-              <div>
-                15. Upload or transmit (or attempt to upload or to
-                transmit) any material that acts as a passive or active
-                information collection or transmission mechanism, including
-                without limitation, clear graphics interchange formats
-                (&#8220;gifs&#8221;), 1&#215;1 pixels, web bugs, cookies,
-                or other similar devices (sometimes referred to as
-                &#8220;spyware&#8221; or &#8220;passive collection
-                mechanisms&#8221; or &#8220;pcms&#8221;).
-              </div>
-              <div>
-                16. Interfere with, disrupt, or create an undue burden on
-                the Site or the networks or services connected to the Site.
-              </div>
-              <div>
-                17. Harass, annoy, intimidate, or threaten any of our
-                employees or agents engaged in providing any portion of the
-                Site to you.
-              </div>
-              <div>
-                18. Attempt to bypass any measures of the Site designed to
-                prevent or restrict access to the Site, or any portion of
-                the Site.
-              </div>
-              <div>
-                19. Copy or adapt the Site&#8217;s software, including but
-                not limited to Flash, PHP, HTML, JavaScript, or other code.
-              </div>
-              <div>
-                20. Decipher, decompile, disassemble, or reverse engineer
-                any of the software comprising or in any way making up a
-                part of the Site.
-              </div>
-              <div>
-                21. Except as may be the result of standard search engine
-                or Internet browser usage, use, launch, develop, or
-                distribute any automated system, including without
-                limitation, any spider, robot, cheat utility, scraper, or
-                offline reader that accesses the Site, or using or
-                launching any unauthorized script or other software.
-              </div>
-              <div>
-                22. Use a buying agent or purchasing agent to make
-                purchases on the Site.
-              </div>
-              <div>
-                23. Make any unauthorized use of the Site, including
-                collecting usernames and/or email addresses of users by
-                electronic or other means for the purpose of sending
-                unsolicited email, or creating user accounts by automated
-                means or under false pretenses.
-              </div>
-              <div>
-                24. Use the Site as part of any effort to compete with us
-                or otherwise use the Site and/or the Content for any
-                revenue-generating endeavor or commercial enterprise.
-              </div>
-              <div />
-              <div>
-                <a name="_zbbv9tgty199" />
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                USER GENERATED CONTRIBUTIONS
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                The Site does not offer users to submit or post content. We
-                may provide you with the opportunity to create, submit,
-                post, display, transmit, perform, publish, distribute, or
-                broadcast content and materials to us or on the Site,
-                including but not limited to text, writings, video, audio,
-                photographs, graphics, comments, suggestions, or personal
-                information or other material (collectively,
-                "Contributions"). Contributions may be viewable by other
-                users of the Site and through third-party websites. As
-                such, any Contributions you transmit may be treated in
-                accordance with the Site Privacy Policy. When you create or
-                make available any Contributions, you thereby represent and
-                warrant that:
-              </div>
-            </div>
-            <div>
-              <div>
-                <br />
-              </div>
-              <div>
-                <div>
-                  1. The creation, distribution, transmission, public
-                  display, or performance, and the accessing,
-                  downloading, or copying of your Contributions do not
-                  and will not infringe the proprietary rights, including
-                  but not limited to the copyright, patent, trademark,
-                  trade secret, or moral rights of any third party.
-                  <br />
-                  2. You are the creator and owner of or have the
-                  necessary licenses, rights, consents, releases, and
-                  permissions to use and to authorize us, the Site, and
-                  other users of the Site to use your Contributions in
-                  any manner contemplated by the Site and these Terms of
-                  Use.
-                  <br />
-                  3. You have the written consent, release, and/or
-                  permission of each and every identifiable individual
-                  person in your Contributions to use the name or
-                  likeness of each and every such identifiable individual
-                  person to enable inclusion and use of your
-                  Contributions in any manner contemplated by the Site
-                  and these Terms of Use.
-                  <br />
-                  4. Your Contributions are not false, inaccurate, or
-                  misleading.
-                  <br />
-                  5. Your Contributions are not unsolicited or
-                  unauthorized advertising, promotional materials,
-                  pyramid schemes, chain letters, spam, mass mailings, or
-                  other forms of solicitation.
-                  <br />
-                  6. Your Contributions are not obscene, lewd,
-                  lascivious, filthy, violent, harassing, libelous,
-                  slanderous, or otherwise objectionable (as determined
-                  by us).
-                  <br />
-                  7. Your Contributions do not ridicule, mock, disparage,
-                  intimidate, or abuse anyone.
-                  <br />
-                  8. Your Contributions do not advocate the violent
-                  overthrow of any government or incite, encourage, or
-                  threaten physical harm against another.
-                  <br />
-                  9. Your Contributions do not violate any applicable
-                  law, regulation, or rule.
-                  <br />
-                  10. Your Contributions do not violate the privacy or
-                  publicity rights of any third party.
-                  <br />
-                  11. Your Contributions do not contain any material that
-                  solicits personal information from anyone under the age
-                  of 18 or exploits people under the age of 18 in a
-                  sexual or violent manner.
-                  <br />
-                  12. Your Contributions do not violate any applicable
-                  law concerning child pornography, or otherwise intended
-                  to protect the health or well-being of minors;
-                  <br />
-                  13. Your Contributions do not include any offensive
-                  comments that are connected to race, national origin,
-                  gender, sexual preference, or physical handicap.
-                  <br />
-                  14. Your Contributions do not otherwise violate, or
-                  link to material that violates, any provision of these
-                  Terms of Use, or any applicable law or regulation.
-                </div>
-                <div>
-                  <br />
-                </div>
-              </div>
-              <div>
-                Any use of the Site in violation of the foregoing violates
-                these Terms of Use and may result in, among other things,
-                termination or suspension of your rights to use the Site.
-              </div>
-            </div>
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            <div>
-              <div>
-                CONTRIBUTION LICENSE
-              </div>
-              <div />
-            </div>
-            <div>
-              <br />
-            </div>
-          </div>
-          <div>
-            You and the Site agree that we may access, store, process, and use
-            any information and personal data that you provide following the
-            terms of the Privacy Policy and your choices (including settings).
-            <br />
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            By submitting suggestions or other feedback regarding the Site, you
-            agree that we can use and share such feedback for any purpose
-            without compensation to you.
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            We do not assert any ownership over your Contributions. You retain
-            full ownership of all of your Contributions and any intellectual
-            property rights or other proprietary rights associated with your
-            Contributions. We are not liable for any statements or
-            representations in your Contributions provided by you in any area
-            on the Site. You are solely responsible for your Contributions to
-            the Site and you expressly agree to exonerate us from any and all
-            responsibility and to refrain from any legal action against us
-            regarding your Contributions.
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            <br />
-          </div>
-          <div />
-        </div>
-        <div>
-          <a name="_6nl7u6ag6use" />
-        </div>
-        <div>
-          MOBILE APPLICATION LICENSE
-        </div>
-      </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            <a name="_4p0l96wt8fhj" />
-          </div>
-          <div>
-            Use License
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            If you access the Site via a mobile application, then we grant you a
-            revocable, non-exclusive, non-transferable, limited right to install
-            and use the mobile application on wireless electronic devices owned or
-            controlled by you, and to access and use the mobile application on such
-            devices strictly in accordance with the terms and conditions of this
-            mobile application license contained in these Terms of Use. You shall
-            not: (1) decompile, reverse engineer, disassemble, attempt to derive
-            the source code of, or decrypt the application; (2) make any
-            modification, adaptation, improvement, enhancement, translation, or
-            derivative work from the application; (3) violate any applicable laws,
-            rules, or regulations in connection with your access or use of the
-            application; (4) remove, alter, or obscure any proprietary notice
-            (including any notice of copyright or trademark) posted by us or the
-            licensors of the application; (5) use the application for any revenue
-            generating endeavor, commercial enterprise, or other purpose for which
-            it is not designed or intended; (6) make the application available over
-            a network or other environment permitting access or use by multiple
-            devices or users at the same time; (7) use the application for creating
-            a product, service, or software that is, directly or indirectly,
-            competitive with or in any way a substitute for the application; (8)
-            use the application to send automated queries to any website or to send
-            any unsolicited commercial e-mail; or (9) use any proprietary
-            information or any of our interfaces or our other intellectual property
-            in the design, development, manufacture, licensing, or distribution of
-            any applications, accessories, or devices for use with the application.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            <a name="_vzf0b5xscg" />
-            Apple and Android Devices
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            The following terms apply when you use a mobile application obtained
-            from either the Apple Store or Google Play (each an &#8220;App
-            Distributor&#8221;) to access the Site: (1) the license granted to you
-            for our mobile application is limited to a non-transferable license to
-            use the application on a device that utilizes the Apple iOS or Android
-            operating systems, as applicable, and in accordance with the usage
-            rules set forth in the applicable App Distributor&#8217;s terms of
-            service; (2) we are responsible for providing any maintenance and
-            support services with respect to the mobile application as specified in
-            the terms and conditions of this mobile application license contained
-            in these Terms of Use or as otherwise required under applicable law,
-            and you acknowledge that each App Distributor has no obligation
-            whatsoever to furnish any maintenance and support services with respect
-            to the mobile application; (3) in the event of any failure of the
-            mobile application to conform to any applicable warranty, you may
-            notify the applicable App Distributor, and the App Distributor, in
-            accordance with its terms and policies, may refund the purchase price,
-            if any, paid for the mobile application, and to the maximum extent
-            permitted by applicable law, the App Distributor will have no other
-            warranty obligation whatsoever with respect to the mobile application;
-            (4) you represent and warrant that (i) you are not located in a country
-            that is subject to a U.S. government embargo, or that has been
-            designated by the U.S. government as a &#8220;terrorist
-            supporting&#8221; country and (ii) you are not listed on any U.S.
-            government list of prohibited or restricted parties; (5) you must
-            comply with applicable third-party terms of agreement when using the
-            mobile application, e.g., if you have a VoIP application, then you must
-            not be in violation of their wireless data service agreement when using
-            the mobile application; and (6) you acknowledge and agree that the App
-            Distributors are third-party beneficiaries of the terms and conditions
-            in this mobile application license contained in these Terms of Use, and
-            that each App Distributor will have the right (and will be deemed to
-            have accepted the right) to enforce the terms and conditions in this
-            mobile application license contained in these Terms of Use against you
-            as a third-party beneficiary thereof.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center" />
-        <div align="center" />
-        <div>
-          SOCIAL MEDIA
-        </div>
-        <div>
-          <br />
-        </div>
-        <div>
-          As part of the functionality of the Site, you may link your account with
-          online accounts you have with third-party service providers (each such
-          account, a &#8220;Third-Party Account&#8221;) by either: (1) providing your
-          Third-Party Account login information through the Site; or (2) allowing us
-          to access your Third-Party Account, as is permitted under the applicable
-          terms and conditions that govern your use of each Third-Party Account. You
-          represent and warrant that you are entitled to disclose your Third-Party
-          Account login information to us and/or grant us access to your Third-Party
-          Account, without breach by you of any of the terms and conditions that
-          govern your use of the applicable Third-Party Account, and without
-          obligating us to pay any fees or making us subject to any usage limitations
-          imposed by the third-party service provider of the Third-Party Account. By
-          granting us access to any Third-Party Accounts, you understand that (1) we
-          may access, make available, and store (if applicable) any content that you
-          have provided to and stored in your Third-Party Account (the &#8220;Social
-          Network Content&#8221;) so that it is available on and through the Site via
-          your account, including without limitation any friend lists and (2) we may
-          submit to and receive from your Third-Party Account additional information
-          to the extent you are notified when you link your account with the
-          Third-Party Account. Depending on the Third-Party Accounts you choose and
-          subject to the privacy settings that you have set in such Third-Party
-          Accounts, personally identifiable information that you post to your
-          Third-Party Accounts may be available on and through your account on the
-          Site. Please note that if a Third-Party Account or associated service
-          becomes unavailable or our access to such Third Party Account is terminated
-          by the third-party service provider, then Social Network Content may no
-          longer be available on and through the Site. You will have the ability to
-          disable the connection between your account on the Site and your
-          Third-Party Accounts at any time. PLEASE NOTE THAT YOUR RELATIONSHIP WITH
-          THE THIRD-PARTY SERVICE PROVIDERS ASSOCIATED WITH YOUR THIRD-PARTY ACCOUNTS
-          IS GOVERNED SOLELY BY YOUR AGREEMENT(S) WITH SUCH THIRD-PARTY SERVICE
-          PROVIDERS. We make no effort to review any Social Network Content for any
-          purpose, including but not limited to, for accuracy, legality, or
-          non-infringement, and we are not responsible for any Social Network
-          Content. You acknowledge and agree that we may access your email address
-          book associated with a Third-Party Account and your contacts list stored on
-          your mobile device or tablet computer solely for purposes of identifying
-          and informing you of those contacts who have also registered to use the
-          Site. You can deactivate the connection between the Site and your
-          Third-Party Account by contacting us using the contact information below or
-          through your account settings (if applicable). We will attempt to delete
-          any information stored on our servers that was obtained through such
-          Third-Party Account, except the username and profile picture that become
-          associated with your account.
-        </div>
-        <div>
-          <br />
-        </div>
-        <div>
-          <br />
-        </div>
-        <div align="center" />
-        <div>
-          SUBMISSIONS
-        </div>
-        <div>
-          <br />
-        </div>
-        <div>
-          You acknowledge and agree that any questions, comments, suggestions, ideas,
-          feedback, or other information regarding the Site ("Submissions") provided
-          by you to us are non-confidential and shall become our sole property. We
-          shall own exclusive rights, including all intellectual property rights, and
-          shall be entitled to the unrestricted use and dissemination of these
-          Submissions for any lawful purpose, commercial or otherwise, without
-          acknowledgment or compensation to you. You hereby waive all moral rights to
-          any such Submissions, and you hereby warrant that any such Submissions are
-          original with you or that you have the right to submit such Submissions.
-          You agree there shall be no recourse against us for any alleged or actual
-          infringement or misappropriation of any proprietary right in your
-          Submissions.
-        </div>
-        <div>
-          <br />
-        </div>
-        <div>
-          <br />
-        </div>
-        <div align="center" />
-        <div>
-          THIRD-PARTY WEBSITE AND CONTENT
-        </div>
-        <div>
-          <br />
-        </div>
-        <div>
-          The Site may contain (or you may be sent via the Site) links to other
-          websites ("Third-Party Websites") as well as articles, photographs, text,
-          graphics, pictures, designs, music, sound, video, information,
-          applications, software, and other content or items belonging to or
-          originating from third parties ("Third-Party Content"). Such Third-Party
-          Websites and Third-Party Content are not investigated, monitored, or
-          checked for accuracy, appropriateness, or completeness by us, and we are
-          not responsible for any Third-Party Websites accessed through the Site or
-          any Third-Party Content posted on, available through, or installed from the
-          Site, including the content, accuracy, offensiveness, opinions,
-          reliability, privacy practices, or other policies of or contained in the
-          Third-Party Websites or the Third-Party Content. Inclusion of, linking to,
-          or permitting the use or installation of any Third-Party Websites or any
-          Third-Party Content does not imply approval or endorsement thereof by us.
-          If you decide to leave the Site and access the Third-Party Websites or to
-          use or install any Third-Party Content, you do so at your own risk, and you
-          should be aware these Terms of Use no longer govern. You should review the
-          applicable terms and policies, including privacy and data gathering
-          practices, of any website to which you navigate from the Site or relating
-          to any applications you use or install from the Site. Any purchases you
-          make through Third-Party Websites will be through other websites and from
-          other companies, and we take no responsibility whatsoever in relation to
-          such purchases which are exclusively between you and the applicable third
-          party. You agree and acknowledge that we do not endorse the products or
-          services offered on Third-Party Websites and you shall hold us harmless
-          from any harm caused by your purchase of such products or services.
-          Additionally, you shall hold us harmless from any losses sustained by you
-          or harm caused to you relating to or resulting in any way from any
-          Third-Party Content or any contact with Third-Party Websites.
-        </div>
-        <div>
-          <br />
-        </div>
-        <div>
-          <br />
-        </div>
-        <div align="center" />
-        <div align="center">
-          <div>
-            <a name="_29ce8o9pbtmi" />
-          </div>
-          <div>
-            <a name="_wj13r09u8u3u" />
-            SITE MANAGEMENT
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            We reserve the right, but not the obligation, to: (1) monitor the Site
-            for violations of these Terms of Use; (2) take appropriate legal action
-            against anyone who, in our sole discretion, violates the law or these
-            Terms of Use, including without limitation, reporting such user to law
-            enforcement authorities; (3) in our sole discretion and without
-            limitation, refuse, restrict access to, limit the availability of, or
-            disable (to the extent technologically feasible) any of your
-            Contributions or any portion thereof; (4) in our sole discretion and
-            without limitation, notice, or liability, to remove from the Site or
-            otherwise disable all files and content that are excessive in size or
-            are in any way burdensome to our systems; and (5) otherwise manage the
-            Site in a manner designed to protect our rights and property and to
-            facilitate the proper functioning of the Site.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div>
-            <br />
-          </div>
-          <div>
-            <a name="_jugvcvcw0oj9" />
-          </div>
-          <div>
-            PRIVACY POLICY
-          </div>
-          <div>
-            <br />
-          </div>
-          <div>
-            We care about data privacy and security. Please review our Privacy
-            Policy: downatdart.me/legal. By using the Site, you agree to be bound
-            by our Privacy Policy, which is incorporated into these Terms of Use.
-            Please be advised the Site is hosted in the United States. If you
-            access the Site from any other region of the world with laws or other
-            requirements governing personal data collection, use, or disclosure
-            that differ from applicable laws in the United States, then through
-            your continued use of the Site, you are transferring your data to the
-            United States, and you agree to have your data transferred to and
-            processed in the United States.
-          </div>
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <br />
-        </div>
-        <div align="center">
-          <div />
-          <div>
-            <a name="_n081pott8yce" />
-          </div>
-          <div>
-            <a name="_sg28ikxq3swh" />
-          </div>
-          <div align="center">
-            <div>
-              <a name="_k3mndam4w6w1" />
-              TERM AND TERMINATION
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              These Terms of Use shall remain in full force and effect while you
-              use the Site. WITHOUT LIMITING ANY OTHER PROVISION OF THESE TERMS
-              OF USE, WE RESERVE THE RIGHT TO, IN OUR SOLE DISCRETION AND WITHOUT
-              NOTICE OR LIABILITY, DENY ACCESS TO AND USE OF THE SITE (INCLUDING
-              BLOCKING CERTAIN IP ADDRESSES), TO ANY PERSON FOR ANY REASON OR FOR
-              NO REASON, INCLUDING WITHOUT LIMITATION FOR BREACH OF ANY
-              REPRESENTATION, WARRANTY, OR COVENANT CONTAINED IN THESE TERMS OF
-              USE OR OF ANY APPLICABLE LAW OR REGULATION. WE MAY TERMINATE YOUR
-              USE OR PARTICIPATION IN THE SITE OR DELETE YOUR ACCOUNT AND ANY
-              CONTENT OR INFORMATION THAT YOU POSTED AT ANY TIME, WITHOUT
-              WARNING, IN OUR SOLE DISCRETION.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              If we terminate or suspend your account for any reason, you are
-              prohibited from registering and creating a new account under your
-              name, a fake or borrowed name, or the name of any third party, even
-              if you may be acting on behalf of the third party. In addition to
-              terminating or suspending your account, we reserve the right to
-              take appropriate legal action, including without limitation
-              pursuing civil, criminal, and injunctive redress.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_e2dep1hfgltt" />
-              MODIFICATIONS AND INTERRUPTIONS
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              We reserve the right to change, modify, or remove the contents of
-              the Site at any time or for any reason at our sole discretion
-              without notice. However, we have no obligation to update any
-              information on our Site. We also reserve the right to modify or
-              discontinue all or part of the Site without notice at any time. We
-              will not be liable to you or any third party for any modification,
-              price change, suspension, or discontinuance of the Site.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              We cannot guarantee the Site will be available at all times. We may
-              experience hardware, software, or other problems or need to perform
-              maintenance related to the Site, resulting in interruptions,
-              delays, or errors. We reserve the right to change, revise, update,
-              suspend, discontinue, or otherwise modify the Site at any time or
-              for any reason without notice to you. You agree that we have no
-              liability whatsoever for any loss, damage, or inconvenience caused
-              by your inability to access or use the Site during any downtime or
-              discontinuance of the Site. Nothing in these Terms of Use will be
-              construed to obligate us to maintain and support the Site or to
-              supply any corrections, updates, or releases in connection
-              therewith.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_p6vbf8atcwhs" />
-              GOVERNING LAW
-            </div>
-            <div>
-              <br />
-            </div>
-            <div />
-            <div>
-              These Terms of Use and your use of the Site are governed by and
-              construed in accordance with the laws of the State of North
-              Carolina applicable to agreements made and to be entirely performed
-              within the State of North Carolina, without regard to its conflict
-              of law principles.
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_v5i5tjw62cyw" />
-              DISPUTE RESOLUTION
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <div>
-                Informal Negotiations
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                To expedite resolution and control the cost of any dispute,
-                controversy, or claim related to these Terms of Use (each a
-                "Dispute" and collectively, the &#8220;Disputes&#8221;) brought
-                by either you or us (individually, a &#8220;Party&#8221; and
-                collectively, the &#8220;Parties&#8221;), the Parties agree to
-                first attempt to negotiate any Dispute (except those Disputes
-                expressly provided below) informally for at least thirty (30)
-                days before initiating arbitration. Such informal negotiations
-                commence upon written notice from one Party to the other Party.
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                <a name="_p2enhfx5sw3j" />
-                Binding Arbitration
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                If the Parties are unable to resolve a Dispute through informal
-                negotiations, the Dispute (except those Disputes expressly
-                excluded below) will be finally and exclusively resolved
-                through binding arbitration. YOU UNDERSTAND THAT WITHOUT THIS
-                PROVISION, YOU WOULD HAVE THE RIGHT TO SUE IN COURT AND HAVE A
-                JURY TRIAL. The arbitration shall be commenced and conducted
-                under the Commercial Arbitration Rules of the American
-                Arbitration Association ("AAA") and, where appropriate, the
-                AAA&#8217;s Supplementary Procedures for Consumer Related
-                Disputes ("AAA Consumer Rules"), both of which are available at
-                the AAA website:
-                <a
-                  href="http://www.adr.org/"
-                  rel="noopener noreferrer"
-                  target="_blank"
-                >
-                  www.adr.org
-                </a>
-                . Your arbitration fees and your share of arbitrator
-                compensation shall be governed by the AAA Consumer Rules and,
-                where appropriate, limited by the AAA Consumer Rules. The
-                arbitration may be conducted in person, through the submission
-                of documents, by phone, or online. The arbitrator will make a
-                decision in writing, but need not provide a statement of
-                reasons unless requested by either Party. The arbitrator must
-                follow applicable law, and any award may be challenged if the
-                arbitrator fails to do so. Except where otherwise required by
-                the applicable AAA rules or applicable law, the arbitration
-                will take place in Wake, North Carolina. Except as otherwise
-                provided herein, the Parties may litigate in court to compel
-                arbitration, stay proceedings pending arbitration, or to
-                confirm, modify, vacate, or enter judgment on the award entered
-                by the arbitrator.
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                If for any reason, a Dispute proceeds in court rather than
-                arbitration, the Dispute shall be commenced or prosecuted in
-                the state and federal courts located in Wake, North Carolina,
-                and the Parties hereby consent to, and waive all defenses of
-                lack of personal jurisdiction, and forum non conveniens with
-                respect to venue and jurisdiction in such state and federal
-                courts. Application of the United Nations Convention on
-                Contracts for the International Sale of Goods and the the
-                Uniform Computer Information Transaction Act (UCITA) is
-                excluded from these Terms of Use.
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                In no event shall any Dispute brought by either Party related
-                in any way to the Site be commenced more than one (1) years
-                after the cause of action arose. If this provision is found to
-                be illegal or unenforceable, then neither Party will elect to
-                arbitrate any Dispute falling within that portion of this
-                provision found to be illegal or unenforceable, and such
-                Dispute shall be decided by a court of competent jurisdiction
-                within the courts listed for jurisdiction above, and the
-                Parties agree to submit to the personal jurisdiction of that
-                court.
-              </div>
-              <div>
-                <a name="_xl7dsk6ndrpp" />
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                <a name="_inlv5c77dhih" />
-                Restrictions
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                The Parties agree that any arbitration shall be limited to the
-                Dispute between the Parties individually. To the full extent
-                permitted by law, (a) no arbitration shall be joined with any
-                other proceeding; (b) there is no right or authority for any
-                Dispute to be arbitrated on a class-action basis or to utilize
-                class action procedures; and (c) there is no right or authority
-                for any Dispute to be brought in a purported representative
-                capacity on behalf of the general public or any other persons.
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                <a name="_mdjlt1af25uq" />
-                Exceptions to Informal Negotiations and Arbitration
-              </div>
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <div>
-                The Parties agree that the following Disputes are not subject
-                to the above provisions concerning informal negotiations and
-                binding arbitration: (a) any Disputes seeking to enforce or
-                protect, or concerning the validity of, any of the intellectual
-                property rights of a Party; (b) any Dispute related to, or
-                arising from, allegations of theft, piracy, invasion of
-                privacy, or unauthorized use; and (c) any claim for injunctive
-                relief. If this provision is found to be illegal or
-                unenforceable, then neither Party will elect to arbitrate any
-                Dispute falling within that portion of this provision found to
-                be illegal or unenforceable and such Dispute shall be decided
-                by a court of competent jurisdiction within the courts listed
-                for jurisdiction above, and the Parties agree to submit to the
-                personal jurisdiction of that court.
-              </div>
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_mjgzo07ttzx5" />
-              CORRECTIONS
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              There may be information on the Site that contains typographical
-              errors, inaccuracies, or omissions, including descriptions,
-              pricing, availability, and various other information. We reserve
-              the right to correct any errors, inaccuracies, or omissions and to
-              change or update the information on the Site at any time, without
-              prior notice.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_gvi74blrahf9" />
-              DISCLAIMER
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              THE SITE IS PROVIDED ON AN AS-IS AND AS-AVAILABLE BASIS. YOU AGREE
-              THAT YOUR USE OF THE SITE AND OUR SERVICES WILL BE AT YOUR SOLE
-              RISK. TO THE FULLEST EXTENT PERMITTED BY LAW, WE DISCLAIM ALL
-              WARRANTIES, EXPRESS OR IMPLIED, IN CONNECTION WITH THE SITE AND
-              YOUR USE THEREOF, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
-              WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE,
-              AND NON-INFRINGEMENT. WE MAKE NO WARRANTIES OR REPRESENTATIONS
-              ABOUT THE ACCURACY OR COMPLETENESS OF THE SITE&#8217;S CONTENT OR
-              THE CONTENT OF ANY WEBSITES LINKED TO THE SITE AND WE WILL ASSUME
-              NO LIABILITY OR RESPONSIBILITY FOR ANY (1) ERRORS, MISTAKES, OR
-              INACCURACIES OF CONTENT AND MATERIALS, (2) PERSONAL INJURY OR
-              PROPERTY DAMAGE, OF ANY NATURE WHATSOEVER, RESULTING FROM YOUR
-              ACCESS TO AND USE OF THE SITE, (3) ANY UNAUTHORIZED ACCESS TO OR
-              USE OF OUR SECURE SERVERS AND/OR ANY AND ALL PERSONAL INFORMATION
-              AND/OR FINANCIAL INFORMATION STORED THEREIN, (4) ANY INTERRUPTION
-              OR CESSATION OF TRANSMISSION TO OR FROM THE SITE, (5) ANY BUGS,
-              VIRUSES, TROJAN HORSES, OR THE LIKE WHICH MAY BE TRANSMITTED TO OR
-              THROUGH THE SITE BY ANY THIRD PARTY, AND/OR (6) ANY ERRORS OR
-              OMISSIONS IN ANY CONTENT AND MATERIALS OR FOR ANY LOSS OR DAMAGE OF
-              ANY KIND INCURRED AS A RESULT OF THE USE OF ANY CONTENT POSTED,
-              TRANSMITTED, OR OTHERWISE MADE AVAILABLE VIA THE SITE. WE DO NOT
-              WARRANT, ENDORSE, GUARANTEE, OR ASSUME RESPONSIBILITY FOR ANY
-              PRODUCT OR SERVICE ADVERTISED OR OFFERED BY A THIRD PARTY THROUGH
-              THE SITE, ANY HYPERLINKED WEBSITE, OR ANY WEBSITE OR MOBILE
-              APPLICATION FEATURED IN ANY BANNER OR OTHER ADVERTISING, AND WE
-              WILL NOT BE A PARTY TO OR IN ANY WAY BE RESPONSIBLE FOR MONITORING
-              ANY TRANSACTION BETWEEN YOU AND ANY THIRD-PARTY PROVIDERS OF
-              PRODUCTS OR SERVICES. AS WITH THE PURCHASE OF A PRODUCT OR SERVICE
-              THROUGH ANY MEDIUM OR IN ANY ENVIRONMENT, YOU SHOULD USE YOUR BEST
-              JUDGMENT AND EXERCISE CAUTION WHERE APPROPRIATE.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_4pjah3d0455q" />
-              LIMITATIONS OF LIABILITY
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              IN NO EVENT WILL WE OR OUR DIRECTORS, EMPLOYEES, OR AGENTS BE
-              LIABLE TO YOU OR ANY THIRD PARTY FOR ANY DIRECT, INDIRECT,
-              CONSEQUENTIAL, EXEMPLARY, INCIDENTAL, SPECIAL, OR PUNITIVE DAMAGES,
-              INCLUDING LOST PROFIT, LOST REVENUE, LOSS OF DATA, OR OTHER DAMAGES
-              ARISING FROM YOUR USE OF THE SITE, EVEN IF WE HAVE BEEN ADVISED OF
-              THE POSSIBILITY OF SUCH DAMAGES. NOTWITHSTANDING ANYTHING TO THE
-              CONTRARY CONTAINED HEREIN, OUR LIABILITY TO YOU FOR ANY CAUSE
-              WHATSOEVER AND REGARDLESS OF THE FORM OF THE ACTION, WILL AT ALL
-              TIMES BE LIMITED TO THE AMOUNT PAID, IF ANY, BY YOU TO US. CERTAIN
-              US STATE LAWS AND INTERNATIONAL LAWS DO NOT ALLOW LIMITATIONS ON
-              IMPLIED WARRANTIES OR THE EXCLUSION OR LIMITATION OF CERTAIN
-              DAMAGES. IF THESE LAWS APPLY TO YOU, SOME OR ALL OF THE ABOVE
-              DISCLAIMERS OR LIMITATIONS MAY NOT APPLY TO YOU, AND YOU MAY HAVE
-              ADDITIONAL RIGHTS.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_k5ap68aj1dd4" />
-              INDEMNIFICATION
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              You agree to defend, indemnify, and hold us harmless, including our
-              subsidiaries, affiliates, and all of our respective officers,
-              agents, partners, and employees, from and against any loss, damage,
-              liability, claim, or demand, including reasonable attorneys&#8217;
-              fees and expenses, made by any third party due to or arising out
-              of: (1) use of the Site; (2) breach of these Terms of Use; (3) any
-              breach of your representations and warranties set forth in these
-              Terms of Use; (4) your violation of the rights of a third party,
-              including but not limited to intellectual property rights; or (5)
-              any overt harmful act toward any other user of the Site with whom
-              you connected via the Site. Notwithstanding the foregoing, we
-              reserve the right, at your expense, to assume the exclusive defense
-              and control of any matter for which you are required to indemnify
-              us, and you agree to cooperate, at your expense, with our defense
-              of such claims. We will use reasonable efforts to notify you of any
-              such claim, action, or proceeding which is subject to this
-              indemnification upon becoming aware of it.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_ftgg17oha0ep" />
-              USER DATA
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              We will maintain certain data that you transmit to the Site for the
-              purpose of managing the performance of the Site, as well as data
-              relating to your use of the Site. Although we perform regular
-              routine backups of data, you are solely responsible for all data
-              that you transmit or that relates to any activity you have
-              undertaken using the Site. You agree that we shall have no
-              liability to you for any loss or corruption of any such data, and
-              you hereby waive any right of action against us arising from any
-              such loss or corruption of such data.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_dkovrslqodui" />
-              ELECTRONIC COMMUNICATIONS, TRANSACTIONS, AND SIGNATURES
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              Visiting the Site, sending us emails, and completing online forms
-              constitute electronic communications. You consent to receive
-              electronic communications, and you agree that all agreements,
-              notices, disclosures, and other communications we provide to you
-              electronically, via email and on the Site, satisfy any legal
-              requirement that such communication be in writing. YOU HEREBY AGREE
-              TO THE USE OF ELECTRONIC SIGNATURES, CONTRACTS, ORDERS, AND OTHER
-              RECORDS, AND TO ELECTRONIC DELIVERY OF NOTICES, POLICIES, AND
-              RECORDS OF TRANSACTIONS INITIATED OR COMPLETED BY US OR VIA THE
-              SITE. You hereby waive any rights or requirements under any
-              statutes, regulations, rules, ordinances, or other laws in any
-              jurisdiction which require an original signature or delivery or
-              retention of non-electronic records, or to payments or the granting
-              of credits by any means other than electronic means.
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <br />
-            </div>
-            <div />
-            <div>
-              <a name="_cem9cu2usl7k" />
-              CALIFORNIA USERS AND RESIDENTS
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              If any complaint with us is not satisfactorily resolved, you can
-              contact the Complaint Assistance Unit of the Division of Consumer
-              Services of the California Department of Consumer Affairs in
-              writing at 1625 North Market Blvd., Suite N 112, Sacramento,
-              California 95834 or by telephone at (800) 952-5210 or (916)
-              445-1254.
-            </div>
-            <div>
-              <br />
-            </div>
-            <div>
-              <br />
-            </div>
-            <div />
-            <div>
-              <a name="_d4jvmcnxg0wt" />
-              MISCELLANEOUS
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              These Terms of Use and any policies or operating rules posted by us
-              on the Site or in respect to the Site constitute the entire
-              agreement and understanding between you and us. Our failure to
-              exercise or enforce any right or provision of these Terms of Use
-              shall not operate as a waiver of such right or provision. These
-              Terms of Use operate to the fullest extent permissible by law. We
-              may assign any or all of our rights and obligations to others at
-              any time. We shall not be responsible or liable for any loss,
-              damage, delay, or failure to act caused by any cause beyond our
-              reasonable control. If any provision or part of a provision of
-              these Terms of Use is determined to be unlawful, void, or
-              unenforceable, that provision or part of the provision is deemed
-              severable from these Terms of Use and does not affect the validity
-              and enforceability of any remaining provisions. There is no joint
-              venture, partnership, employment or agency relationship created
-              between you and us as a result of these Terms of Use or use of the
-              Site. You agree that these Terms of Use will not be construed
-              against us by virtue of having drafted them. You hereby waive any
-              and all defenses you may have based on the electronic form of these
-              Terms of Use and the lack of signing by the parties hereto to
-              execute these Terms of Use.
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              <br />
-            </div>
-            <div>
-              <a name="_t4pq5cwn486q" />
-              CONTACT US
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              In order to resolve a complaint regarding the Site or to receive
-              further information regarding use of the Site, please contact us
-              at:
-            </div>
-          </div>
-          <div align="center">
-            <br />
-          </div>
-          <div align="center">
-            <div>
-              Who's Down LLC
-            </div>
-            <div>
-              601 Manassas Gap Place
-            </div>
-            <div>
-              Cary, NC 27519-9613
-            </div>
-            <div>
-              United States
-            </div>
-            <div>
-              Phone: 9197419655
-            </div>
-            <div>
-              bhatt_arjun@yahoo.com
-            </div>
-            <div>
-              <br />
-            </div>
-          </div>
-        </div></div>
-      <h1>Privacy Policy for Who's Down</h1>
-
-      <p>At Who's Down, accessible from downatdart.me, one of our main priorities is the privacy of our visitors. This Privacy Policy document contains types of information that is collected and recorded by Who's Down and how we use it.</p>
-
-      <p>If you have additional questions or require more information about our Privacy Policy, do not hesitate to contact us.</p>
-
-      <p>This Privacy Policy applies only to our online activities and is valid for visitors to our website with regards to the information that they shared and/or collect in Who's Down. This policy is not applicable to any information collected offline or via channels other than this website.</p>
-
-      <h2>Consent</h2>
-
-      <p>By using our website, you hereby consent to our Privacy Policy and agree to its terms.</p>
-
-      <h2>Information we collect</h2>
-
-      <p>The personal information that you are asked to provide, and the reasons why you are asked to provide it, will be made clear to you at the point we ask you to provide your personal information.</p>
-      <p>If you contact us directly, we may receive additional information about you such as your name, email address, phone number, the contents of the message and/or attachments you may send us, and any other information you may choose to provide.</p>
-      <p>When you register for an Account, we may ask for your contact information, including items such as name, company name, address, email address, and telephone number.</p>
-
-      <h2>How we use your information</h2>
-
-      <p>We use the information we collect in various ways, including to:</p>
-
-      <ul>
-        <li>Provide, operate, and maintain our webste</li>
-        <li>Improve, personalize, and expand our webste</li>
-        <li>Understand and analyze how you use our webste</li>
-        <li>Develop new products, services, features, and functionality</li>
-        <li>Communicate with you, either directly or through one of our partners, including for customer service, to provide you with updates and other information relating to the webste, and for marketing and promotional purposes</li>
-        <li>Send you emails</li>
-        <li>Find and prevent fraud</li>
-      </ul>
-
-      <h2>Log Files</h2>
-
-      <p>Who's Down follows a standard procedure of using log files. These files log visitors when they visit websites. All hosting companies do this and a part of hosting services' analytics. The information collected by log files include internet protocol (IP) addresses, browser type, Internet Service Provider (ISP), date and time stamp, referring/exit pages, and possibly the number of clicks. These are not linked to any information that is personally identifiable. The purpose of the information is for analyzing trends, administering the site, tracking users' movement on the website, and gathering demographic information. Our Privacy Policy was created with the help of the <a href="https://www.privacypolicygenerator.info">Privacy Policy Generator</a> and the <a href="https://www.privacypolicyonline.com/privacy-policy-generator/">Online Privacy Policy Generator</a>.</p>
-
-
-      <h2>Advertising Partners Privacy Policies</h2>
-
-      <p>You may consult this list to find the Privacy Policy for each of the advertising partners of Who's Down.</p>
-
-      <p>Third-party ad servers or ad networks uses technologies like cookies, JavaScript, or Web Beacons that are used in their respective advertisements and links that appear on Who's Down, which are sent directly to users' browser. They automatically receive your IP address when this occurs. These technologies are used to measure the effectiveness of their advertising campaigns and/or to personalize the advertising content that you see on websites that you visit.</p>
-
-      <p>Note that Who's Down has no access to or control over these cookies that are used by third-party advertisers.</p>
-
-      <h2>Third Party Privacy Policies</h2>
-
-      <p>Who's Down's Privacy Policy does not apply to other advertisers or websites. Thus, we are advising you to consult the respective Privacy Policies of these third-party ad servers for more detailed information. It may include their practices and instructions about how to opt-out of certain options. </p>
-
-      <p>You can choose to disable cookies through your individual browser options. To know more detailed information about cookie management with specific web browsers, it can be found at the browsers' respective websites.</p>
-
-      <h2>CCPA Privacy Rights (Do Not Sell My Personal Information)</h2>
-
-      <p>Under the CCPA, among other rights, California consumers have the right to:</p>
-      <p>Request that a business that collects a consumer's personal data disclose the categories and specific pieces of personal data that a business has collected about consumers.</p>
-      <p>Request that a business delete any personal data about the consumer that a business has collected.</p>
-      <p>Request that a business that sells a consumer's personal data, not sell the consumer's personal data.</p>
-      <p>If you make a request, we have one month to respond to you. If you would like to exercise any of these rights, please contact us.</p>
-
-      <h2>GDPR Data Protection Rights</h2>
-
-      <p>We would like to make sure you are fully aware of all of your data protection rights. Every user is entitled to the following:</p>
-      <p>The right to access – You have the right to request copies of your personal data. We may charge you a small fee for this service.</p>
-      <p>The right to rectification – You have the right to request that we correct any information you believe is inaccurate. You also have the right to request that we complete the information you believe is incomplete.</p>
-      <p>The right to erasure – You have the right to request that we erase your personal data, under certain conditions.</p>
-      <p>The right to restrict processing – You have the right to request that we restrict the processing of your personal data, under certain conditions.</p>
-      <p>The right to object to processing – You have the right to object to our processing of your personal data, under certain conditions.</p>
-      <p>The right to data portability – You have the right to request that we transfer the data that we have collected to another organization, or directly to you, under certain conditions.</p>
-      <p>If you make a request, we have one month to respond to you. If you would like to exercise any of these rights, please contact us.</p>
-
-      <h2>Children's Information</h2>
-      <p>Another part of our priority is adding protection for children while using the internet. We encourage parents and guardians to observe, participate in, and/or monitor and guide their online activity.</p>
-      <p>Who's Down does not knowingly collect any Personal Identifiable Information from children under the age of 13. If you think that your child provided this kind of information on our website, we strongly encourage you to contact us immediately and we will do our best efforts to promptly remove such information from our records.</p>
-    </div>
-  );
-};
-
 const App = (props) => {
-  // don't touch this
   return (
     <div>
       <Router>
         <div>
           <Nav />
           <Switch>
-            <Route exact path="/" component={Welcome} />
-            <Route path="/whatsup" component={About} />
+            <Route exact path="/" component={Landing} />
+            <Route path="/whatsup" component={Explore} />
             <Route path="/create" component={Create} />
             <Route path="/legal" component={Legal} />
-
+            <Route path="/about" component={About} />
             <Route component={FallBack} />
           </Switch>
+          <Footer />
         </div>
       </Router>
     </div>
