+++ src/style.scss
@@ -25,12 +25,13 @@ $fontfamily: "Montserrat", sans-serif;
   font-size: 36pt;
   text-align: center;
 }
+
 .navBar{
   background-color: white;
   color:#ff5722;
   height: 100%;
   padding-top: 25px;
-  padding-bottom: 100px;
+  padding-bottom: 25px;
 }
 
 .titleBox{
@@ -77,7 +78,7 @@ $fontfamily: "Montserrat", sans-serif;
   align-items: center;
   text-align: center;
   .reasonBox{
-    width: 350px;
+    width: 33%;
     color:orange;
     flex-direction: column;
     justify-content: space-between;
@@ -87,7 +88,7 @@ $fontfamily: "Montserrat", sans-serif;
     &:active{
       color: green;
     }
-  }  
+  }
 }
 
 
@@ -176,7 +177,7 @@ $fontfamily: "Montserrat", sans-serif;
 }
 
 .active {
-  color: pink;
+  color: #ff5722;
 }
 
 // .container {
@@ -190,5 +191,156 @@ $fontfamily: "Montserrat", sans-serif;
 
 .item {
   order: 1;
-  
+}
+
+// Nav CSS
+
+.navBar {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: no-wrap;
+  justify-content: space-evenly;
+  flex-flow: row;
+  list-style-type: none;
+
+  li {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+  }
+}
+
+.navItem {
+  text-decoration: none;
+  border-radius: 4px;
+
+  &:hover {
+  color: orange;
+  fontSize: 20px;
+  text-decoration: none;
+  }
+}
+
+.donateItem {
+  display: inline-block;
+  border-radius: 4px;
+  background-color: #f4511e;
+  border: none;
+  color: #FFFFFF;
+  text-align: center;
+  font-size: 14px;
+  padding: 10px;
+  width: 100px;
+  transition: all 0.5s;
+  cursor: pointer;
+  margin: 5px;
+  text-decoration: none;
+
+  &:hover {
+    text-decoration: none;
+    color: #FFFFFF;
+  }
+}
+
+.donateItem span {
+  cursor: pointer;
+  display: inline-block;
+  position: relative;
+  transition: 0.5s;
+}
+
+.donateItem span:after {
+  content: '\00bb';
+  position: absolute;
+  opacity: 0;
+  top: 0;
+  right: -20px;
+  transition: 0.5s;
+}
+
+.donateItem:hover span {
+  padding-right: 25px;
+  text-decoration: none;
+}
+
+.donateItem:hover span:after {
+  opacity: 1;
+  right: 0;
+  text-decoration: none;
+}
+
+// Footer CSS
+.footerContainer {
+  padding-right: 40px;
+  padding-left: 40px;
+}
+
+.footerBar {
+  border-top: 1px solid lightgrey;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: no-wrap;
+  justify-content: space-around;
+  list-style-type: none;
+  margin-top: 40px;
+  padding-top: 20px;
+  padding-bottom: 8px;
+  padding-left: 0px;
+
+  li {
+    flex-basis: 33.333333%;
+    flex-direction: row;
+    display: flex;
+    justify-content: center;
+  }
+}
+
+.footerItem {
+  text-decoration: none;
+
+  &:hover {
+  color: orange;
+  fontSize: 20px;
+  text-decoration: none;
+  }
+}
+
+.footerLogo {
+  max-width: 10%;
+  max-height: 10%;
+}
+
+
+//Landing CSS
+
+.styleObj{
+    font-size: 25;
+    font-weight: bold;
+    color: #B5D8CF;
+    padding-top: 100px;
+    background-color: black;
+}
+
+.boldText{
+    font-family: Montserrat, sans-serif;
+    font-size: 20;
+    font-weight: bold;
+    color: black;
+    justify-content: center;
+}
+
+.blueHeaderStyle{
+    font-family: Montserrat, sans-serif;
+    font-size: 25;
+    font-weight: bold;
+    color: #005975;
+    padding-top: 100px;
+}
+
+.blueBodyStyle{
+    font-family: 'Montserrat, sans-serif';
+    font-size: 12;
+    font-weight: 'extra-light';
+    color: '#005975';
 }
